created factory module and tested

+++ x/factory/types/message_mint_and_send_tokens.go
@@ -9,41 +9,40 @@ const TypeMsgMintAndSendTokens = "mint_and_send_tokens"
 
 var _ sdk.Msg = &MsgMintAndSendTokens{}
 
-func NewMsgMintAndSendTokens(owner string, denom string, amount int32, recipient string) *MsgMintAndSendTokens {
-  return &MsgMintAndSendTokens{
-		Owner: owner,
-    Denom: denom,
-    Amount: amount,
-    Recipient: recipient,
+func NewMsgMintAndSendTokens(owner string, denom string, amount uint64, recipient string) *MsgMintAndSendTokens {
+	return &MsgMintAndSendTokens{
+		Owner:     owner,
+		Denom:     denom,
+		Amount:    amount,
+		Recipient: recipient,
 	}
 }
 
 func (msg *MsgMintAndSendTokens) Route() string {
-  return RouterKey
+	return RouterKey
 }
 
 func (msg *MsgMintAndSendTokens) Type() string {
-  return TypeMsgMintAndSendTokens
+	return TypeMsgMintAndSendTokens
 }
 
 func (msg *MsgMintAndSendTokens) GetSigners() []sdk.AccAddress {
-  owner, err := sdk.AccAddressFromBech32(msg.Owner)
-  if err != nil {
-    panic(err)
-  }
-  return []sdk.AccAddress{owner}
+	owner, err := sdk.AccAddressFromBech32(msg.Owner)
+	if err != nil {
+		panic(err)
+	}
+	return []sdk.AccAddress{owner}
 }
 
 func (msg *MsgMintAndSendTokens) GetSignBytes() []byte {
-  bz := ModuleCdc.MustMarshalJSON(msg)
-  return sdk.MustSortJSON(bz)
+	bz := ModuleCdc.MustMarshalJSON(msg)
+	return sdk.MustSortJSON(bz)
 }
 
 func (msg *MsgMintAndSendTokens) ValidateBasic() error {
-  _, err := sdk.AccAddressFromBech32(msg.Owner)
-  	if err != nil {
-  		return sdkerrors.Wrapf(sdkerrors.ErrInvalidAddress, "invalid owner address (%s)", err)
-  	}
-  return nil
+	_, err := sdk.AccAddressFromBech32(msg.Owner)
+	if err != nil {
+		return sdkerrors.Wrapf(sdkerrors.ErrInvalidAddress, "invalid owner address (%s)", err)
+	}
+	return nil
 }
-
